+++ transports/quic/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.10.3
-- Allow configuring MTU discovery with `MtuDiscoveryConfig`.
+- Allow configuring MTU discovery upper bound.
+  See [PR 5386](https://github.com/libp2p/rust-libp2p/pull/5386).
 
 ## 0.10.2